requirements.txt
Flask>=2.0
Flask-WTF>=1.0
Flask-Login>=0.6
Flask-SQLAlchemy>=3.0
WTForms>=3.0
SQLAlchemy>=2.0
Werkzeug>=2.0
python-dotenv>=1.0
gunicorn>=20.0
requests>=2.0
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib

Flask-Migrate
psycopg2-binary
pytz
openai>=1.0
pytest-mock
